+++ Loogn.OrmLite/OrmLiteWriteApiTrans.cs
@@ -52,13 +52,13 @@ namespace Loogn.OrmLite
             return SqlHelper.ExecuteNonQuery(dbTrans, commandType, commandText, ORM.DictionaryToParams(parameters));
         }
 
-        public static long Insert<T>(this SqlTransaction dbTrans, T obj, bool selectIdentity = false)
+        public static int Insert<T>(this SqlTransaction dbTrans, T obj, bool selectIdentity = false)
         {
             var tuple = SqlCmd.Insert<T>(obj, selectIdentity);
             if (selectIdentity)
             {
                 var identity = ExecuteScalar(dbTrans, CommandType.Text, tuple.Item1, tuple.Item2);
-                return Convert.ToInt64(identity);
+                return Convert.ToInt32(identity);
             }
             else
             {
@@ -67,13 +67,13 @@ namespace Loogn.OrmLite
             }
         }
 
-        public static long Insert(this SqlTransaction dbTrans, string table, Dictionary<string, object> fields, bool selectIdentity = false)
+        public static int Insert(this SqlTransaction dbTrans, string table, Dictionary<string, object> fields, bool selectIdentity = false)
         {
             var tuple = SqlCmd.Insert(table, fields, selectIdentity);
             if (selectIdentity)
             {
                 var identity = ExecuteScalar(dbTrans, CommandType.Text, tuple.Item1, tuple.Item2);
-                return Convert.ToInt64(identity);
+                return Convert.ToInt32(identity);
             }
             else
             {
@@ -82,13 +82,13 @@ namespace Loogn.OrmLite
             }
         }
 
-        public static long Insert(this SqlTransaction dbTrans, string table, object anonType, bool selectIdentity = false)
+        public static int Insert(this SqlTransaction dbTrans, string table, object anonType, bool selectIdentity = false)
         {
             var tuple = SqlCmd.Insert(table, anonType, selectIdentity);
             if (selectIdentity)
             {
                 var identity = ExecuteScalar(dbTrans, CommandType.Text, tuple.Item1, tuple.Item2);
-                return Convert.ToInt64(identity);
+                return Convert.ToInt32(identity);
             }
             else
             {
